requirements.txt
boto==2.45.0
bz2file==0.98
Cython==0.25.2
gensim==0.13.4.1
mecab-python3==0.7
numpy==1.12.0
requests==2.13.0
scipy==0.18.1
six==1.10.0
smart-open==1.3.5
